backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pandas
urllib3
gunicorn
